requirements.txt
python-dotenv
langgraph
langchain
ipython
langchain-openai
langchain-community
langchain-core
langchain-chroma
chromadb
pypdf
openai
beautifulsoup4
lxml
html5lib
trafilatura
unstructured
unstructured[pdf]
pdfminer.six
streamlit
fastapi
uvicorn 